consolidated the Context providers Removed double checks for data availability
